Export java functions to their own module
